+++ www/src/components/Plural.js
@@ -5,6 +5,7 @@ import { StripeProvider } from 'react-stripe-elements'
 import ApplicationLayout from './layout/ApplicationLayout'
 import BreadcrumbProvider from './Breadcrumbs'
 import Chart from './repos/Chart'
+import CloudShell from './shell/CloudShell'
 import EditUser from './users/EditUser'
 import Marketplace from './marketplace/Marketplace'
 import OAuthCallback from './shell/OAuthCallback'
@@ -19,32 +20,31 @@ import RepositoryPackagesHelm from './repository/RepositoryPackagesHelm'
 import RepositoryPackagesTerraform from './repository/RepositoryPackagesTerraform'
 import RepositoryTests from './repository/RepositoryTests'
 import Terraform from './repos/Terraform'
-import CloudShell from './shell/CloudShell'
+import { AccessTokens } from './profile/AccessTokens'
+import { Account } from './account/Account'
+import { AccountAttributes } from './account/AccountAttributes'
+import { AuditChloro, AuditDirectory, Audits, LoginAudits } from './account/Audits'
+import { Clusters } from './clusters/Clusters'
 import { CurrentUserContext, PluralConfigurationContext, PluralProvider } from './login/CurrentUser'
 import { DeviceLoginNotif } from './users/DeviceLoginNotif'
 import { Docker, DockerRepository } from './repos/Docker'
-import { EditAccount } from './accounts/EditAccount'
-import { IntegrationPage } from './repos/Integrations'
-import { OauthCreator } from './integrations/OauthCreator'
-import { UpgradeQueue } from './upgrades/UpgradeQueue'
-import { UpgradeQueues } from './upgrades/UpgradeQueues'
-import { VerifyEmailConfirmed } from './users/EmailConfirmation'
-import { MyProfile } from './profile/MyProfile'
-import { Me } from './profile/Me'
-import { Security } from './profile/Security'
-import { AccessTokens } from './profile/AccessTokens'
-import { PublicKeys } from './profile/PublicKeys'
-import { EabCredentials } from './profile/EabCredentials'
-import { Account } from './account/Account'
-import { Users } from './account/Users'
-import { ServiceAccounts } from './account/ServiceAccounts'
-import { Groups } from './account/Groups'
-import { Roles } from './account/Roles'
 import { Domains } from './account/Domains'
-import { AccountAttributes } from './account/AccountAttributes'
-import { Clusters } from './upgrades/Clusters'
-import { AuditChloro, AuditDirectory, Audits, LoginAudits } from './account/Audits'
+import { EabCredentials } from './profile/EabCredentials'
+import { EditAccount } from './accounts/EditAccount'
+import { Groups } from './account/Groups'
+import { IntegrationPage } from './repos/Integrations'
+import { Me } from './profile/Me'
+import { MyProfile } from './profile/MyProfile'
 import { OIDCProvider } from './repository/OIDCProvider'
+import { OauthCreator } from './integrations/OauthCreator'
+import { PublicKeys } from './profile/PublicKeys'
+import { Roles } from './account/Roles'
+import { Security } from './profile/Security'
+import { ServiceAccounts } from './account/ServiceAccounts'
+import { UpgradeQueue } from './clusters/UpgradeQueue'
+import { UpgradeQueues } from './clusters/UpgradeQueues'
+import { Users } from './account/Users'
+import { VerifyEmailConfirmed } from './users/EmailConfirmation'
 
 function EditBilling(props) {
   return (
